msm: kgsl: Fix sysfs deadlock in kgsl
If kgsl mem entries are being accessed in sysfs and if putting
back the mementry refcount triggers a free of the process private,
we have a mutex deadlock. Because freeing the process private triggers
removal of the sysfs directories from within a thread that is accessing
the sysfs files. Fix this by taking an extra refcount on the process
private and then putting it back in a deferred manner.

Change-Id: I7db0e6144cabec2a86df9afbc500cd0ba3af0291
Signed-off-by: Harshdeep Dhatt <hdhatt@codeaurora.org>
Signed-off-by: prabhatkrmishra <mprabhat774@gmail.com>