arm64: vdso32: Convert to modern assembler annotations
In an effort to clarify and simplify the annotation of assembly
functions new macros have been introduced. These replace ENTRY and
ENDPROC with two different annotations for normal functions and those
with unusual calling conventions. Use these for the compat VDSO,
allowing us to drop the custom ARM_ENTRY() and ARM_ENDPROC() macros.

Change-Id: Iea6c8890c734014d41b0e5600f8f511e8043a3bd
Signed-off-by: 0xA11CE613 <mprabhat774@gmail.com>